% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.polaaar.EML.data}
\alias{get.polaaar.EML.data}
\title{extract information from an onine EML document}
\usage{
get.polaaar.EML.data(EML.url)
}
\arguments{
\item{EML.url}{the URL to the EML document}
}
\value{
a list with the data needed for the POLAAAR database. This includes: abstract, start_date, end_date, bounding_box, is_public and associated_references
}
\description{
get.polaaar.EML.data extracts the information needed for the POLAAAR database from a given onine EML document
}
\author{
Maxime Sweetlove ccBY 4.0 2019
}
