% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_POLAAARimport.R
\name{metadata.to.polaaar}
\alias{metadata.to.polaaar}
\title{format data to upload to the POLAAAR database}
\usage{
metadata.to.polaaar(metadata.object, EML.url = NA, user_ID = "",
  dest.dir = getwd(), create.dir = FALSE)
}
\arguments{
\item{metadata.object}{a metadata.MIxS, DwC.event or a DwC.occurrence class object. The object to be formatted to fit the polaaar database schema}

\item{EML.url}{a named vector. URL to the EML (IPT) associated with the dataset, written as a named vector, e.g. c(project1="http://url1", project2="http://url1"). The names in the vector must correspond to the different projects listed under "project_name" in the metadata object. If only one link is provided and is unnamed, than this will automatically be the parent project to all the samples in the metadata object.}

\item{user_ID}{ID of the user that send the data}

\item{dest.dir}{a character vector. The complete path to the directory where the CSV files should be written to. Default is the working diectory. If create.dir is TRUE, the given directory will be created if it does not exist.}

\item{create.dir}{boolean. If TRUE, the directory path in the variable dest.dir will be created if it does not already exist. Default FALSE}
}
\value{
data is written to POLAAAR (currently as different CSV files that correspond to the POLAAAR database tables)
}
\description{
metadata.to.polaaar writes a metadata.MIxS, DwC.event or a DwC.occurrence class object object to the polaaar database
}
\details{
POLAAAR, which is short for Polar ’Omics Links in Arctic-Antarctic-Alpine (A3) Research, is an online thematic portal that focusses molecular biodiversity data resources from polar and alpine regions. Our approach is based on enriching public datasets with the associated metadata and environmental information. These data are made publicly accessible through POLA3R and are also linked to the associated publications and the sequences on INSDC. To allow interoperability with other systems, the portal is designed to operate between different data archiving standards, such as the Minimum Information on any (x) Sequence (MIxS) as well as DarwinCore. Datasets that are listed on POLA3R are also registered on GBIF to increase their discoverability.
}
\author{
Maxime Sweetlove ccBY 4.0 2019
}
