% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess.polaaar.DwC.core}
\alias{preprocess.polaaar.DwC.core}
\title{preprocess DarwinCore data for POLAAAR}
\usage{
preprocess.polaaar.DwC.core(dataset)
}
\arguments{
\item{dataset}{a data.frame formated with DarwinCore terms}
}
\value{
a data.frame that has been prepared for metadata.to.polaaar()
}
\description{
preprocess DarwinCore data for POLAAAR
}
\author{
Maxime Sweetlove ccBY 4.0 2019
}
