% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess.polaaar.MIxS}
\alias{preprocess.polaaar.MIxS}
\title{preprocess metadata.MIxS for POLAAAR}
\usage{
preprocess.polaaar.MIxS(metadata.object)
}
\arguments{
\item{metadata.object}{a metadata.MIxS object}

\item{metadata.object}{a metadata.MIxS object}
}
\value{
a metadata.MIxS object that has been prepared for metadata.to.polaaar()

a metadata.MIxS object that has been prepared for metadata.to.polaaar()
}
\description{
preprocess metadata.MIxS for POLAAAR
}
\author{
Maxime Sweetlove ccBY 4.0 2019
}
